+++ storage/src/persistent/commit_log.rs
@@ -1,8 +1,7 @@
-use crate::persistent::commit_log::error::TezedgeCommitLogError;
-use crate::persistent::commit_log::reader::Reader;
-use crate::persistent::commit_log::writer::Writer;
+// Copyright (c) SimpleStaking and Tezedge Contributors
+// SPDX-License-Identifier: MIT
+
 use std::collections::HashMap;
-use std::convert::TryInto;
 use std::path::{Path, PathBuf};
 use std::sync::{Arc, RwLock};
 use std::{fmt, io};
@@ -10,50 +9,13 @@ use std::{fmt, io};
 use failure::Fail;
 use serde::{Deserialize, Serialize};
 
+use crate::commit_log::error::TezedgeCommitLogError;
+use crate::commit_log::CommitLog;
 use crate::persistent::codec::{Decoder, Encoder, SchemaError};
 use crate::persistent::schema::{CommitLogDescriptor, CommitLogSchema};
 use crate::persistent::BincodeEncoded;
 
 pub type CommitLogRef = Arc<RwLock<CommitLog>>;
-pub mod error;
-mod reader;
-mod writer;
-
-const INDEX_FILE_NAME: &str = "table.index";
-const DATA_FILE_NAME: &str = "table.data";
-
-const TH_LENGTH: usize = 24;
-
-pub type Message = Vec<u8>;
-pub struct MessageSet {
-    cursor: usize,
-    indexes: Vec<Index>,
-    acc: usize,
-    buf: Vec<u8>,
-}
-
-impl MessageSet {
-    fn new(indexes: Vec<Index>, buf: Vec<u8>) -> Self {
-        Self {
-            cursor: 0,
-            indexes,
-            acc: 0,
-            buf,
-        }
-    }
-}
-
-impl Iterator for MessageSet {
-    type Item = Message;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        let index = self.indexes.get(self.cursor)?;
-        let data = &self.buf[self.acc..(self.acc + index.uncompressed_data_length as usize)];
-        self.acc += index.uncompressed_data_length as usize;
-        self.cursor += 1;
-        Some(data.to_vec())
-    }
-}
 
 /// Possible errors for commit log
 #[derive(Debug, Fail)]
@@ -70,130 +32,6 @@ pub enum CommitLogError {
     CorruptData,
     #[fail(display = "Tezedge CommitLog error: {:#?}", error)]
     TezedgeCommitLogError { error: TezedgeCommitLogError },
-}
-
-#[derive(Debug, Copy, Clone)]
-pub struct Index {
-    pub position: u64,
-    pub compressed_data_length: u64,
-    pub uncompressed_data_length: u64,
-}
-
-impl Index {
-    fn from_buf(buf: &[u8]) -> Result<Self, TezedgeCommitLogError> {
-        if buf.len() != TH_LENGTH {
-            return Err(TezedgeCommitLogError::IndexLengthError);
-        }
-        let mut buf = buf.to_vec();
-        let position_raw_bytes: Vec<_> = buf.drain(..8).collect();
-        let compressed_data_length_raw_bytes: Vec<_> = buf.drain(..8).collect();
-        let uncompressed_data_length_raw_bytes: Vec<_> = buf.drain(..).collect();
-        let position = u64::from_be_bytes(
-            position_raw_bytes
-                .as_slice()
-                .try_into()
-                .map_err(|_| TezedgeCommitLogError::TryFromSliceError)?,
-        );
-        let compressed_data_length = u64::from_be_bytes(
-            compressed_data_length_raw_bytes
-                .as_slice()
-                .try_into()
-                .map_err(|_| TezedgeCommitLogError::TryFromSliceError)?,
-        );
-        let uncompressed_data_length = u64::from_be_bytes(
-            uncompressed_data_length_raw_bytes
-                .as_slice()
-                .try_into()
-                .map_err(|_| TezedgeCommitLogError::TryFromSliceError)?,
-        );
-        Ok(Self {
-            position,
-            compressed_data_length,
-            uncompressed_data_length,
-        })
-    }
-    fn new(position: u64, compressed_data_length: u64, uncompressed_data_length: u64) -> Self {
-        Self {
-            position,
-            compressed_data_length,
-            uncompressed_data_length,
-        }
-    }
-    fn to_vec(&self) -> Vec<u8> {
-        let mut buf = vec![];
-        for b in &self.position.to_be_bytes() {
-            buf.push(*b)
-        }
-        for b in &self.compressed_data_length.to_be_bytes() {
-            buf.push(*b)
-        }
-        for b in &self.uncompressed_data_length.to_be_bytes() {
-            buf.push(*b)
-        }
-        return buf;
-    }
-}
-
-pub struct CommitLog {
-    path: PathBuf,
-}
-
-impl CommitLog {
-    pub fn new<P: AsRef<Path>>(log_dir: P) -> Self {
-        Self {
-            path: log_dir.as_ref().to_path_buf(),
-        }
-    }
-    #[inline]
-    pub fn append_msg<B: AsRef<[u8]>>(&mut self, payload: B) -> Result<u64, TezedgeCommitLogError> {
-        let mut writer = Writer::new(&self.path)?;
-        let offset = writer.write(payload.as_ref())?;
-        Ok(offset)
-    }
-
-    #[inline]
-    pub fn read(&self, from: usize, limit: usize) -> Result<MessageSet, TezedgeCommitLogError> {
-        let mut reader = Reader::new(&self.path)?;
-        return reader.range(from, limit);
-    }
-
-    pub fn iter(&mut self) -> Result<CommitLogIterator, TezedgeCommitLogError> {
-        let reader = Reader::new(&self.path)?;
-        Ok(CommitLogIterator::new(reader))
-    }
-
-    pub fn flush(&mut self) -> Result<(), TezedgeCommitLogError> {
-        let mut writer = Writer::new(&self.path)?;
-        return writer.flush();
-    }
-}
-
-pub struct CommitLogIterator {
-    cursor: u64,
-    reader: Reader,
-}
-
-impl CommitLogIterator {
-    fn new(reader: Reader) -> Self {
-        Self { cursor: 0, reader }
-    }
-}
-
-impl Iterator for CommitLogIterator {
-    type Item = Message;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        if self.reader.indexes.get(self.cursor as usize).is_none() {
-            return None;
-        }
-        match self.reader.read_at(self.cursor as usize) {
-            Ok(m) => {
-                self.cursor += 1;
-                Some(m)
-            }
-            Err(_) => None,
-        }
-    }
 }
 
 impl From<SchemaError> for CommitLogError {
@@ -364,7 +202,7 @@ impl CommitLogs {
         if !Path::new(&path).exists() {
             std::fs::create_dir_all(&path)?;
         }
-        let log = CommitLog::new(path);
+        let log = CommitLog::new(path)?;
 
         let mut commit_log_map = self.commit_log_map.write().unwrap();
         commit_log_map.insert(name.into(), Arc::new(RwLock::new(log)));
